- move ClubDetailed classes to own package
