+++ app/src/main/AndroidManifest.xml
@@ -12,7 +12,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".ui.userviews.club.ClubDetailedActivity"
+            android:name=".ui.userviews.club.detailed.ClubDetailedActivity"
             android:theme="@style/AppTheme.NoActionBar"/>
         <activity
             android:name=".ui.login.LoginActivity"
