+++ app/src/main/java/no/ntnu/klubbhuset/ui/userviews/club/detailed/ClubDetailedActivity.java
@@ -1,32 +1,16 @@
-package no.ntnu.klubbhuset.ui.userviews.club;
+package no.ntnu.klubbhuset.ui.userviews.club.detailed;
 
-import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
 
-import com.google.android.material.floatingactionbutton.FloatingActionButton;
-import com.google.android.material.snackbar.Snackbar;
-
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.appcompat.widget.Toolbar;
 import androidx.fragment.app.Fragment;
-import androidx.fragment.app.FragmentActivity;
 import androidx.fragment.app.FragmentTransaction;
-import androidx.lifecycle.ViewModel;
 import androidx.lifecycle.ViewModelProviders;
-
-import android.util.AttributeSet;
-import android.view.View;
-import android.widget.Button;
-import android.widget.TextView;
-
-import java.util.Objects;
 
 import no.ntnu.klubbhuset.R;
 import no.ntnu.klubbhuset.data.model.Club;
 import no.ntnu.klubbhuset.data.model.Member;
-
-import static java.security.AccessController.getContext;
 
 public class ClubDetailedActivity extends AppCompatActivity implements ClubDetailedFragment.onMembershipStatusChangedListener {
 
@@ -53,15 +37,18 @@ public class ClubDetailedActivity extends AppCompatActivity implements ClubDetai
         FragmentTransaction transaction = getSupportFragmentManager().beginTransaction()
                 .replace(R.id.club_detailed_fragment_container, newFragment);
         transaction.commit();
-
-
     }
 
 
+    /**
+     * listens for changes in membership state, e.g when user joins an org,
+     * or membership status gets fetched.
+     * @param member
+     */
     @Override
     public void onMembershipStatusChanged(Member member) {
-
         if (member != null) {
+            // replace non-member view with member-view
             Fragment newFragment = ClubDetailedMember.newInstance(member);
             FragmentTransaction transaction = getSupportFragmentManager().beginTransaction()
                     .replace(R.id.club_detailed_fragment_container, newFragment);
