+++ app/src/main/res/layout/activity_club_detailed.xml
@@ -4,11 +4,11 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.userviews.club.ClubDetailedActivity">
+    tools:context=".ui.userviews.club.detailed.ClubDetailedActivity">
 
     <fragment
         android:id="@+id/fragment2"
-        android:name="no.ntnu.klubbhuset.ui.userviews.club.ClubDetailedFragment"
+        android:name="no.ntnu.klubbhuset.ui.userviews.club.detailed.ClubDetailedFragment"
         android:layout_width="match_parent"
         android:layout_height="match_parent" />
 </FrameLayout>
