+++ app/src/main/res/layout/club_detailed_not_member_fragment.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="no.ntnu.klubbhuset.ui.userviews.club.ClubDetailedNotMember">
+    tools:context="no.ntnu.klubbhuset.ui.userviews.club.detailed.ClubDetailedNotMember">
 
     <Button
         android:id="@+id/club_detailed_joinbtn"
