+++ app/src/main/res/layout/content_club_detailed_member.xml
@@ -5,7 +5,7 @@
     android:id="@+id/club_detailed_fragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.userviews.club.ClubDetailedMember">
+    tools:context=".ui.userviews.club.detailed.ClubDetailedMember">
 
     <LinearLayout
         android:layout_width="match_parent"
